requirements.txt
flask~=3.0.2
flask-cors
numpy~=1.26.2
Pillow~=10.2.0
requests~=2.31.0
rembg~=2.0.54
openai~=1.12.0
opencv-python-headless
h11~=0.14.0
typing_extensions~=4.9.0
pip~=23.2.1
distro~=1.9.0
Jinja2~=3.1.3
sniffio~=1.3.0
MarkupSafe~=2.1.5
Werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
httpcore~=1.0.3
idna~=3.6
certifi~=2024.2.2
pytz~=2023.3.post1
tqdm~=4.66.2
httpx~=0.26.0
pydantic~=2.6.1
anyio~=4.3.0
pandas~=2.1.4
urllib3~=2.2.1
pydantic_core~=2.16.2
pymongo~=4.6.1
attrs~=23.2.0
exceptiongroup~=1.2.0
numba~=0.59.0
llvmlite~=0.42.0
setuptools~=58.1.0
scipy~=1.12.0
pooch~=1.8.1
packaging~=23.2
platformdirs~=4.2.0
PyMatting~=1.1.12
onnxruntime~=1.17.0
jsonschema~=4.21.1
mpmath~=1.3.0
sympy~=1.12
imageio~=2.34.0
tifffile~=2024.2.12
scikit-image~=0.22.0
networkx~=3.2.1
referencing~=0.33.0
zipp~=3.17.0
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=23.5.26
python-dotenv~=1.0.1